adding more to the serialization
