+++ twister2/comms/src/java/edu/iu/dsc/tws/comms/dfw/ChannelMessage.java
@@ -23,13 +23,13 @@ public class ChannelMessage {
   /**
    * List of buffers filled with the message
    */
-  private final List<DataBuffer> buffers = new ArrayList<DataBuffer>();
+  protected final List<DataBuffer> buffers = new ArrayList<DataBuffer>();
 
   /**
    * List of byte arrays which are used to copy data from {@link ChannelMessage#buffers}
    * When the system runs out of receive buffers
    */
-  private final List<DataBuffer> overflowBuffers = new ArrayList<DataBuffer>();
+  protected final List<DataBuffer> overflowBuffers = new ArrayList<DataBuffer>();
 
   /**
    * Keeps the number of references to this message
@@ -55,7 +55,7 @@ public class ChannelMessage {
   /**
    * The message header
    */
-  private MessageHeader header;
+  protected MessageHeader header;
 
   /**
    * Keeps track of whether header of the object contained in the buffer was sent or not.
@@ -72,9 +72,9 @@ public class ChannelMessage {
   private boolean isPartial;
 
   /**
-   * Keep whether the message has been fully built
+   * Keep whether we have all the buffers added
    */
-  private boolean complete = false;
+  protected boolean complete = false;
 
   /**
    * Message type
@@ -90,12 +90,6 @@ public class ChannelMessage {
    * Number of bytes in the header
    */
   private int headerSize;
-
-  /**
-   * The current reading buffer index, this can change from 0 > to 0 when we release the buffers
-   * while we are still reading
-   */
-  private int currentReadingBuffer;
 
   public enum ReceivedState {
     INIT,
@@ -117,35 +111,6 @@ public class ChannelMessage {
    * Keep track weather out count updated
    */
   private boolean outCountUpdated = false;
-
-  /**
-   * Number of objects we have read so far
-   */
-  private int objectsDeserialized = 0;
-
-  /**
-   * The object that is been built
-   */
-  private Object deserializingObject;
-
-  /**
-   * The current buffer read index
-   */
-  private int currentBufferIndex = 0;
-
-  /**
-   * Number of buffers added
-   */
-  private int addedBuffers = 0;
-
-  // the amount of data we have seen for current object
-  private int previousReadForObject = 0;
-
-  // keep track of the current object length
-  private int currentObjectLength = 0;
-
-  // the objects we have in buffers so far
-  private int seenObjects = 0;
 
 
   public ChannelMessage() {
@@ -213,45 +178,6 @@ public class ChannelMessage {
     buffers.add(buffer);
   }
 
-  public boolean addBufferAndCalculate(DataBuffer buffer) {
-    buffers.add(buffer);
-    addedBuffers++;
-
-    int expectedObjects = header.getNumberTuples();
-    int remaining = 0;
-    if (addedBuffers == 1) {
-      currentObjectLength = buffer.getByteBuffer().getInt(16);
-      remaining = buffer.getByteBuffer().remaining() + Integer.BYTES;
-    }
-
-
-    while (remaining > 0) {
-      // need to read this much
-      int moreToReadForCurrentObject = currentObjectLength - previousReadForObject;
-      // amount of data in the buffer
-      if (moreToReadForCurrentObject < remaining) {
-        seenObjects++;
-        remaining = remaining - moreToReadForCurrentObject;
-      } else {
-        previousReadForObject += remaining;
-        break;
-      }
-
-      // if we have seen all, lets break
-      if (expectedObjects == seenObjects) {
-        complete = true;
-        break;
-      }
-
-      // we can read another object
-      if (remaining > Integer.BYTES) {
-        currentObjectLength = buffer.getByteBuffer().getInt();
-        previousReadForObject = 0;
-      }
-    }
-
-    return complete;
-  }
 
   protected void addBuffers(List<DataBuffer> bufferList) {
     buffers.addAll(bufferList);
